refactor(node): split protocol.rs into domain modules
Break the 1485-line monolithic Server impl into protocol/ with thin
trait delegates in mod.rs and domain logic in separate files:
conversation, agent, config, plugin, admin, history.
